.scripts/requirements.txt
qrcode[pil]>=8.2,<9
python-slugify>=8,<9
